letter: wrote you a heartfelt letter and got those cool animations you liked from ruoyu and kristen's wedding website
I guess this is a wrap for 2nd year anniversary gift. I know you've been wanting a website for a while and I hope that this is something you revisit often. Given that it's a website, you can always file some github issues for me if you want new features or just wanna see how this looks :p